Just save a single copy of the fluorescent pixels per frame
This had been kept in CellHulls.greenInd; that is, a full copy per hull.  Now
I'm just using that field as a boolean which is true if there's an intersection
between the fluorescent pixels and the phase segmentation.

The fluorescent pixels where already being kept in FluorData.greenInd, so I'm
just using those to draw the green pixels on the screen in DrawCells.