rewrote drawFluorMarkers() to make logic simpler to understand
Now it breaks down the drawing into 3 parts -- before the first fluor frame,
in the middle of the fluor frames, and from the last fluor frame to the end
of the track.

Aside from that, the big change in this version is that I don't mark the track
as fluorescent until I've seen at least one fluor frame.  The prior version
start the fluoresence marker on the frame *before* the one where it turned on.