create semaphore file to indicate the segmentation is finished
The idea behind this is to prevent a race condition in LEVer where it starts
processing the segmentation file before HematoSeg.exe has finished writing it
all out. So when the thread has written the entire file and closed it, we now
create a new 0-byte semaphore file to indicate that it's OK for LEVer to
process it. In case the segmentation gets run more than once, we delete all of
the semaphore files for this movie in the initialization code, before the
threads start running.