always use the index and size of the second-biggest region
GetLabels() returns a vector of labeled regions. The largest one is the entire
image, and the second-largest one is the region we want to crop. The code was
assuming that the regions were returned in order of size, but we found a movie
where that wasn't the case. So now we look at all the regions, compute all
their areas, and use the one that's second-biggest.