Only look for submarines when alpha == 1
This is because the submarine detection code runs Michel, which can take a
while.  Upstream code can call this function multiple times with different
alphas, but we only want to call the slow Michel code once.

A better solution would probably be to refactor things to make the calling code
smarter, but at least this makes things better for the time being.